examples/AWS_Strands/agent_only/.any_agent/Minimal Strands Agent-docker-context/requirements.txt
# AWS Strands Agent Dependencies
strands-agents
strands-agents-tools
python-dotenv
anthropic

# Framework-specific A2A dependencies
strands-agents[a2a]
uvicorn[standard]
fastapi
a2a-sdk>=0.1.0